- Fixed typo - Moved queries to a different file
